cloud-functions/upload_fhir_data/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudevents==1.4.0
deprecation==2.1.0
Flask==2.1.3
functions-framework==3.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
packaging==21.3
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.11
watchdog==2.1.9
Werkzeug==2.2.0
zipp==3.8.1
